moved PIDController to examples folder
