+++ examples/drone/quadcopter_controller.py
@@ -2,8 +2,21 @@ import torch
 import numpy as np
 from genesis.engine.entities.drone_entity import DroneEntity
 from genesis.utils.geom import quat_to_xyz
-from genesis.utils.pid import PIDController
 
+class PIDController():
+    def __init__(self, kp, ki, kd):
+        self.kp = kp
+        self.ki = ki
+        self.kd = kd
+        self.integral = 0.
+        self.prev_error = 0.
+
+    def update(self, error, dt):
+        self.integral += error * dt
+        derivative = (error - self.prev_error) / dt
+        self.prev_error = error
+
+        return (self.kp*error) + (self.ki*self.integral) + (self.kd*derivative)
 
 class DronePIDController:
     def __init__(self, drone: DroneEntity, dt, base_rpm, pid_params):
