--- genesis/utils/pid.py
@@ -1,16 +1,0 @@
-## generalized PID controller for any classical control
-
-class PIDController():
-    def __init__(self, kp, ki, kd):
-        self.kp = kp
-        self.ki = ki
-        self.kd = kd
-        self.integral = 0.
-        self.prev_error = 0.
-
-    def update(self, error, dt):
-        self.integral += error * dt
-        derivative = (error - self.prev_error) / dt
-        self.prev_error = error
-
-        return (self.kp*error) + (self.ki*self.integral) + (self.kd*derivative)
